+++ src/render.rs
@@ -19,7 +19,7 @@ pub fn render(
 
   for (position, sprite) in (&positions, &sprites).join() {
     let screen_position = Point::new(position.x as i32 + shake.x, position.y as i32 + shake.y);
-    let screen_rect = Rect::from_center(screen_position, sprite.width, sprite.height);
+    let screen_rect = Rect::from_center(screen_position, sprite.width as u32, sprite.height as u32);
     canvas.copy_ex(
       &textures[sprite.position],
       None,
